+++ sdk/core/Azure.Core/tests/HttpPipelineMessageTest.cs
@@ -18,7 +18,7 @@ namespace Azure.Core.Tests
             var requestMock = new Mock<Request>();
             HttpMessage message = new HttpMessage(requestMock.Object, new ResponseClassifier());
             message.Dispose();
-            requestMock.Verify(r=>r.Dispose(), Times.Once);
+            requestMock.Verify(r => r.Dispose(), Times.Once);
         }
 
         [Test]
@@ -29,8 +29,8 @@ namespace Azure.Core.Tests
             HttpMessage message = new HttpMessage(requestMock.Object, new ResponseClassifier());
             message.Response = responseMock.Object;
             message.Dispose();
-            requestMock.Verify(r=>r.Dispose(), Times.Once);
-            responseMock.Verify(r=>r.Dispose(), Times.Once);
+            requestMock.Verify(r => r.Dispose(), Times.Once);
+            responseMock.Verify(r => r.Dispose(), Times.Once);
         }
 
         [Test]