+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/DocumentAnalysisModels/OperationsLiveTests.cs
@@ -32,12 +32,10 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
         public async Task AnalyzeDocumentOperationCanPollFromNewObject()
         {
             var client = CreateDocumentAnalysisClient(out var nonInstrumentedClient);
-            var modelId = Recording.GenerateId();
 
-            await using var _ = await CreateDisposableBuildModelAsync(modelId);
-
-            var uri = DocumentAnalysisTestEnvironment.CreateUri(TestFile.Blank);
-            var operation = await client.StartAnalyzeDocumentFromUriAsync(modelId, uri);
+            var uri = DocumentAnalysisTestEnvironment.CreateUri(TestFile.ReceiptJpg);
+            var operation = await client.StartAnalyzeDocumentFromUriAsync("prebuilt-receipt", uri);
+            Assert.IsNotNull(operation.GetRawResponse());
 
             var sameOperation = InstrumentOperation(new AnalyzeDocumentOperation(operation.Id, nonInstrumentedClient));
             await sameOperation.WaitForCompletionAsync();
@@ -54,12 +52,29 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             var modelId = Recording.GenerateId();
 
             var operation = await client.StartBuildModelAsync(trainingFilesUri, modelId);
+            Assert.IsNotNull(operation.GetRawResponse());
 
             var sameOperation = InstrumentOperation(new BuildModelOperation(operation.Id, nonInstrumentedClient));
             await sameOperation.WaitForCompletionAsync();
 
             Assert.IsTrue(sameOperation.HasValue);
             Assert.AreEqual(modelId, sameOperation.Value.ModelId);
+        }
+
+        [RecordedTest]
+        public async Task BuildModelOperationPercentageCompletedValue()
+        {
+            var client = CreateDocumentModelAdministrationClient(out var nonInstrumentedClient);
+            var trainingFilesUri = new Uri(TestEnvironment.BlobContainerSasUrl);
+            var modelId = Recording.GenerateId();
+
+            var operation = await client.StartBuildModelAsync(trainingFilesUri, modelId);
+            Assert.AreEqual(0, operation.PercentCompleted);
+
+            await operation.WaitForCompletionAsync();
+
+            Assert.IsTrue(operation.HasValue);
+            Assert.AreEqual(100, operation.PercentCompleted);
         }
 
         [RecordedTest]
@@ -74,12 +89,33 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             CopyAuthorization targetAuth = await client.GetCopyAuthorizationAsync(targetModelId);
 
             var operation = await client.StartCopyModelAsync(trainedModel.ModelId, targetAuth);
+            Assert.IsNotNull(operation.GetRawResponse());
 
             var sameOperation = InstrumentOperation(new CopyModelOperation(operation.Id, nonInstrumentedClient));
             await sameOperation.WaitForCompletionAsync();
 
             Assert.IsTrue(sameOperation.HasValue);
             Assert.AreEqual(targetModelId, sameOperation.Value.ModelId);
+        }
+
+        [RecordedTest]
+        public async Task CopyModelOperationPercentageCompletedValue()
+        {
+            var client = CreateDocumentModelAdministrationClient(out var nonInstrumentedClient);
+            var modelId = Recording.GenerateId();
+
+            await using var trainedModel = await CreateDisposableBuildModelAsync(modelId);
+
+            var targetModelId = Recording.GenerateId();
+            CopyAuthorization targetAuth = await client.GetCopyAuthorizationAsync(targetModelId);
+
+            var operation = await client.StartCopyModelAsync(trainedModel.ModelId, targetAuth);
+            Assert.AreEqual(0, operation.PercentCompleted);
+
+            await operation.WaitForCompletionAsync();
+
+            Assert.IsTrue(operation.HasValue);
+            Assert.AreEqual(100, operation.PercentCompleted);
         }
     }
 }