+++ sdk/identity/Azure.Identity/api/Azure.Identity.netstandard2.0.cs
@@ -41,18 +41,6 @@ namespace Azure.Identity
         public AuthorizationCodeCredentialOptions() { }
         public System.Uri RedirectUri { get { throw null; } set { } }
     }
-    public partial class AzureApplicationCredential : Azure.Core.TokenCredential
-    {
-        public AzureApplicationCredential() { }
-        public AzureApplicationCredential(Azure.Identity.AzureApplicationCredentialOptions options) { }
-        public override Azure.Core.AccessToken GetToken(Azure.Core.TokenRequestContext requestContext, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-        public override System.Threading.Tasks.ValueTask<Azure.Core.AccessToken> GetTokenAsync(Azure.Core.TokenRequestContext requestContext, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-    }
-    public partial class AzureApplicationCredentialOptions : Azure.Identity.TokenCredentialOptions
-    {
-        public AzureApplicationCredentialOptions() { }
-        public string ManagedIdentityClientId { get { throw null; } set { } }
-    }
     public static partial class AzureAuthorityHosts
     {
         public static System.Uri AzureChina { get { throw null; } }
@@ -278,7 +266,8 @@ namespace Azure.Identity
     {
         private object _dummy;
         private int _dummyPrimitive;
-        public System.ReadOnlyMemory<byte> CacheBytes { get { throw null; } set { } }
+        public TokenCacheData(System.ReadOnlyMemory<byte> cacheBytes) { throw null; }
+        public System.ReadOnlyMemory<byte> CacheBytes { get { throw null; } }
     }
     public partial class TokenCachePersistenceOptions
     {