+++ sdk/servicebus/Microsoft.Azure.WebJobs.Extensions.ServiceBus/src/Config/ServiceBusOptions.cs
@@ -128,7 +128,7 @@ namespace Microsoft.Azure.WebJobs.ServiceBus
         /// Gets or sets an optional exception handler that will be invoked if an exception occurs while attempting to process
         /// a message. This does not apply for functions that receive a batch of messages.
         /// </summary>
-        public Func<ProcessErrorEventArgs, Task> ExceptionHandler { get; set; }
+        internal Func<ProcessErrorEventArgs, Task> ExceptionHandler { get; set; }
 
         /// <summary>
         /// Gets or sets the maximum number of messages that will be passed to each function call. This only applies for functions that receive
@@ -156,6 +156,12 @@ namespace Microsoft.Azure.WebJobs.ServiceBus
             NullValueHandling = NullValueHandling.Ignore,
             Formatting = Formatting.Indented
         };
+
+        /// <summary>
+        /// <inheritdoc cref="ServiceBusClientOptions.EnableCrossEntityTransactions"/>
+        /// </summary>
+        public bool EnableCrossEntityTransactions { get; set; }
+
 #pragma warning restore AZC0014 // Avoid using banned types in public API
 
         /// <summary>
@@ -185,7 +191,8 @@ namespace Microsoft.Azure.WebJobs.ServiceBus
                 { nameof(MaxConcurrentCalls), MaxConcurrentCalls },
                 { nameof(MaxConcurrentSessions), MaxConcurrentSessions },
                 { nameof(MaxMessageBatchSize), MaxMessageBatchSize },
-                { nameof(SessionIdleTimeout), SessionIdleTimeout.ToString() ?? string.Empty }
+                { nameof(SessionIdleTimeout), SessionIdleTimeout.ToString() ?? string.Empty },
+                { nameof(EnableCrossEntityTransactions), EnableCrossEntityTransactions }
             };
 
             return options.ToString(Formatting.Indented);
@@ -256,7 +263,8 @@ namespace Microsoft.Azure.WebJobs.ServiceBus
             {
                 RetryOptions = ClientRetryOptions,
                 WebProxy = WebProxy,
-                TransportType = TransportType
+                TransportType = TransportType,
+                EnableCrossEntityTransactions = EnableCrossEntityTransactions
             };
     }
 }